requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
beautifulsoup4==4.12.3
blinker==1.7.0
click==8.1.7
Django==5.0
django-db==0.0.7
django-debug-toolbar==4.2.0
django-rq==2.10.1
django-seed==0.3.1
djangorestframework==3.14.0
docutils==0.20.1
Faker==20.1.0
Flask==3.0.1
gevent==23.9.1
greenlet==3.0.3
html5lib==1.1
itsdangerous==2.1.2
Jinja2==3.1.3
lockfile==0.12.2
Logbook==1.7.0.post0
maildump==1.0
MarkupSafe==2.1.4
passlib==1.7.4
psycopg2-binary==2.9.9
PyMySQL==1.1.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-decouple==3.8
pytz==2023.3.post1
redis==5.0.1
rq==1.15.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
toposort==1.10
typing_extensions==4.8.0
webencodings==0.5.1
Werkzeug==3.0.1
zope.event==5.0
zope.interface==6.1
